moved: core files to engine, recipe files to core
Core had very few tasks before and most of the code was generic. Hence, moved this into engine. Now recipe is an example implementation, however, is also core to the rest of the codebase. Hence, moved recipe to core. In future, if there are any other uses, except engine, every other module needs to be modified.