modified: metrics, fail reasons to reflect the modified transfer process
Earlier, the transfer was done passively with the receiving device making decisions on whether the payload is successfully transmitted or not. With the new design, this philosophy is reversed. The device can attempt transfer and the result will be immediately return. This is done at the Tx step and hence all the relevant measures are modified to reflect this.